requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
xlrd==2.0.1
